==> 02_counterfactual/02_malaria_counterfactual_shapefile_launcher.R <==
########################################################################
#Switch shapefile of Oxford counterfactual csvs for malaria 
#######################################################################

#(1) Setup ############################################################
rm(list = ls())

#user arguments
old_shapefile_version <- '2019_02_27' # malaria csvs generated by Oxford using admin2 from this shapefile
new_shapefile_version <- '2019_09_10' # want to convert results to this shapefile

rasterize_input_dat <- FALSE #if true, will read in input csvs from Oxford and rasterize. Set to false if this step has already been completed and you just want to read in the rasterized results.
tridaly_in_date <- '2019_09_18' #run date where summary csvs for Oxford-generated malaria results are saved
tridaly_out_date <- '2019_10_28' #run date to save new results under

pop_release <- '2019_08_29' 

indicator_list <- c('had_malaria_act','had_malaria_irs','had_malaria_itn','had_malaria_all', 'had_malaria_raw')

#(2) Make qsub ############################################################

for (indicator in indicator_list){
  
  mem <- '100G'
  rt <- '08:00:00'
  queue <- 'geospatial.q'
  proj <- 'proj_geo_nodes'
  name <- paste0(indicator, '_shapefile_switch')
  shell <- '<<< FILEPATH REDACTED >>>'
  code <- '<<< FILEPATH REDACTED >>>/02_counterfactual/02_malaria_counterfactual_shapefile.R'
  
  args <- paste(old_shapefile_version,
                new_shapefile_version,
                tridaly_in_date,
                tridaly_out_date,
                pop_release,
                indicator,
                rasterize_input_dat)
  
  qsub <- paste0('qsub -l m_mem_free=', mem,
                 ' -l fthread=1 -l h_rt=', rt, 
                 ' -v sing_image=default -q ', queue, 
                 ' -P ', proj, 
                 ' -N ', name, 
                 ' ', shell, 
                 ' ', code, 
                 ' ', args)
  
  system(qsub)
}
